+++ src/main/scala/edu/rit/cs/dataunitylab/jsonoid/discovery/schemas/PropertySets.scala
@@ -101,12 +101,12 @@ object PropertySets {
   /** A property set with the minimum set of properties to discover structural
     *  information..
     */
-  def MinProperties(implicit p: JsonoidParams): PropertySet = PropertySet(
+  val MinProperties: PropertySet = PropertySet(
     ArraySchema.MinProperties,
     BooleanSchema.MinProperties,
     IntegerSchema.MinProperties,
     NumberSchema.MinProperties,
-    ObjectSchema.MinProperties(p),
+    ObjectSchema.MinProperties,
     StringSchema.MinProperties
   )