+++ src/test/scala/edu/rit/cs/dataunitylab/jsonoid/discovery/schemas/NullSchemaSpec.scala
@@ -3,8 +3,6 @@ package schemas
 
 import org.json4s._
 
-import PropertySets._
-
 class NullSchemaSpec extends UnitSpec {
   behavior of "NullSchema"